# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTube AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research:
    get:
      operationId: listResearch
      tags: [research]
      summary: List all research jobs
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of research jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Research"
    post:
      operationId: createResearch
      tags: [research]
      summary: Start a research analysis on a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchInput"
      responses:
        "201":
          description: Research job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
  /research/{id}:
    get:
      operationId: getResearch
      tags: [research]
      summary: Get a research job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Research"
        "404":
          description: Not found
    delete:
      operationId: deleteResearch
      tags: [research]
      summary: Delete a research job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── CONTENT STUDIO ──
...
  /pipeline:
    get:
      operationId: listPipeline
      tags: [pipeline]
      summary: List pipeline jobs
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Pipeline jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineJob"
    post:
      operationId: createPipelineJob
      tags: [pipeline]
      summary: Start a new pipeline job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineJobInput"
      responses:
        "201":
          description: Pipeline job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
  /pipeline/{id}:
    get:
      operationId: getPipelineJob
      tags: [pipeline]
      summary: Get pipeline job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"
        "404":
          description: Not found
    patch:
      operationId: updatePipelineJob
      tags: [pipeline]
      summary: Update pipeline job status/step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineJobUpdate"
      responses:
        "200":
          description: Updated pipeline job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineJob"

  # ── YOUTUBE STUDIO ──
...
  /drafts:
    get:
      operationId: listDrafts
      tags: [drafts]
      summary: List draft videos
      responses:
        "200":
          description: Drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draft"
    post:
      operationId: createDraft
      tags: [drafts]
      summary: Create a draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
  /drafts/{id}:
    patch:
      operationId: updateDraft
      tags: [drafts]
      summary: Update a draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
    delete:
      operationId: deleteDraft
      tags: [drafts]
      summary: Delete a draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── SCHEDULES ──
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List scheduled videos
      responses:
        "200":
          description: Schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Schedule a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── AUTOMATION ──
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD SCHEMAS ──
...
    Research:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        status:
          type: string
        shouldCreate:
          type: ["boolean", "null"]
        reasoning:
          type: ["string", "null"]
        topicAnalysis:
          type: ["string", "null"]
        audienceAnalysis:
          type: ["string", "null"]
        keywordAnalysis:
          type: ["string", "null"]
        relatedQuestions:
          type: ["string", "null"]
        competitorResearch:
          type: ["string", "null"]
        searchIntent:
          type: ["string", "null"]
        evergreenPotential:
          type: ["number", "null"]
        viralPotential:
          type: ["number", "null"]
        contentGapAnalysis:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, topic, status, createdAt, updatedAt]
    ResearchInput:
      type: object
      properties:
        topic:
          type: string
          minLength: 2
      required: [topic]

    # ── CONTENT SCHEMAS ──
...
    PipelineJob:
      type: object
      properties:
        id:
          type: integer
        topic:
          type: string
        currentStep:
          type: string
        status:
          type: string
        progress:
          type: integer
        steps:
          type: string
        videoId:
          type: ["integer", "null"]
        contentId:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, topic, currentStep, status, progress, steps, createdAt, updatedAt]
    PipelineJobInput:
      type: object
      properties:
        topic:
          type: string
          minLength: 2
        videoType:
          type: string
      required: [topic, videoType]
    PipelineJobUpdate:
      type: object
      properties:
        currentStep:
          type: string
        status:
          type: string
        progress:
          type: integer
        errorMessage:
          type: string

    # ── YOUTUBE STUDIO SCHEMAS ──
...
    Draft:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
        contentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, type, createdAt, updatedAt]
    DraftInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        type:
          type: string
        notes:
          type: ["string", "null"]
        contentId:
          type: ["integer", "null"]
      required: [title, type]
    DraftUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        notes:
          type: string

    # ── SCHEDULE SCHEMAS ──
    Schedule:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
        videoId:
          type: ["integer", "null"]
        contentId:
          type: ["integer", "null"]
        platform:
          type: string
        createdAt:
          type: string
      required: [id, title, scheduledAt, status, platform, createdAt]
    ScheduleInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        scheduledAt:
          type: string
        videoId:
          type: ["integer", "null"]
        contentId:
          type: ["integer", "null"]
        platform:
          type: string
      required: [title, scheduledAt, platform]
    ScheduleUpdate:
      type: object
      properties:
        title:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
        platform:
          type: string

    # ── AUTOMATION SCHEMAS ──
...