# lib/api-zod/src/generated/types/draftUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * YouTube AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface DraftUpdate {
  title?: string;
  type?: string;
  notes?: string;
}