# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * YouTube AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  /** @minLength 2 */
  title: string;
  scheduledAt: string;
  /** @nullable */
  videoId?: number | null;
  /** @nullable */
  contentId?: number | null;
  platform: string;
}