# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * YouTube AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleUpdate {
  title?: string;
  scheduledAt?: string;
  status?: string;
  platform?: string;
}